==> scripts/02_species_occurrences/05_agregate_occurrences_to_3dgrid.Rmd <==
---
title: "05_agregate_occurrences"
author: "Leire Ibaibarriaga & Mireia Valle"
date: "2022-02-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Description

Script to agregate the occurrences into lon, lat and depth grids

# Load libraries

```{r}
library (here) #enable easy file referencing 
library(dplyr) #data cleaning
library(ggplot2) #plotting
library(cowplot) #plotting more than one plot
library(rgeos) #mapping
library(sp) #mapping
library(rgdal) #mapping
library(maptools) #mapping
library(maps)        # some basic country maps
library(mapdata)     # higher resolution maps
library(marmap)      # access global topography data
library(plyr) #mapping
library(dplyr)
library(tidyverse)
library(readxl)
library(RColorBrewer)# plots
library(ggpointdensity)
library(gridExtra)
library(ncdf4) # to read netcdf
library(grid)

theme_set(theme_bw(base_size=14) + theme(plot.title = element_text(hjust = 0.5)))
```

# Load species data -------------------------------------------------------
```{r}
species <- read.csv (here::here ("data","derived_data", "02_species_occurrences","species_list_FINAL_hemispheres_automatic.csv"),  sep = ",")

species <- species %>% 
  add_column(n_grouped = NA)

# Gridded latitude and longitude ------------------------------------------

# Load bathy data from our study area that is the one used to resample the others
load(here::here("data", "derived_data","variables", "BATHY", "bathy_atl.RData"))

# transform to data frame
bathy_df <- raster::as.data.frame(bathy_atl,xy=TRUE)

# Grid for latitude: 

# sort unique values of latitude
lat_grid <- sort(unique(bathy_df$y))

# mid-points between the unique values
lat_int <- (lat_grid[-1]+lat_grid[-length(lat_grid)])/2

# add extremes to the breakpoints
lat_brks <- c( lat_grid[1]-(lat_int[1]-lat_grid[1]), 
               lat_int, 
               lat_grid[length(lat_grid)]+(lat_int[length(lat_int)]-lat_grid[length(lat_int)]))

# sort unique values of longitude
lon_grid <- sort(unique(bathy_df$x))

# mid-points between the unique values
lon_int <- (lon_grid[-1]+lon_grid[-length(lon_grid)])/2

# add extremes to the breakpoints
lon_brks <- c( lon_grid[1]-(lon_int[1]-lon_grid[1]), 
               lon_int, 
               lon_grid[length(lon_grid)]+(lon_int[length(lon_int)]-lon_grid[length(lon_int)]))


# Gridded depth -----------------------------------------------------------

# load 3D variable to extract depth values

nc_data <- nc_open(here::here("data", "raw_data", "variables", "TEMP", "woa18_decav81B0_t00_04.nc"))

# sort unique values of depth

depth_grid <- ncvar_get(nc_data, "depth")
depth_grid <- depth_grid[1:48] # up to 1050 m depth

# mid-points between the unique values

depth_int <- (depth_grid[-1]+depth_grid[-length(depth_grid)])/2

# add extremes to the breakpoints

depth_brks <- c( -01, 
                 depth_int, 
                 depth_grid[length(depth_grid)]+(depth_int[length(depth_int)]-depth_grid[length(depth_int)]))


# Basic map for plotting --------------------------------------------------

# basic map data

global <- map_data("worldHires")

# get bathymetry data

# bathy <- getNOAA.bathy(lon1=-83,lon2=20,lat1=-90,lat2=90, resolution = 4, keep=FALSE, antimeridian=FALSE)
# bathy.df <- fortify(bathy)

# basic ggplot

p0 <- ggplot() +
  # geom_contour(data=bathy.df, aes(x,y,z=z), breaks=c(-100, -200), col="grey")+
  annotation_map(map=global, fill="grey")+
  coord_sf(xlim=c(-90,50), ylim=c(-90,90))+
  xlab("")+
  ylab("")
```

# Grouped species data ----------------------------------------------------

# Loop to group data for each species according to gridded lon, lat and depth values:
```{r}

for (i in 1:nrow(species)){
  # species, i is the row number and species$SP[i] is the species data name (which is not always paste0("SP",i) )
  
  species[i,]
  
  # Loading data SPi (e.g. SP1, SP2, ... SP76)
  
  load(here::here("data","derived_data", "02_species_occurrences", "02_species_occurrences_no_outliers_automatic", paste0(species$SP[i],"new.RData")))
  
  # Give a generic name
  
  SP <- get(species$SP[i])
  
  # calculate the corresponding gridded lon, lat and depth values
  
  SP <- SP %>% 
    mutate(
      lon_grid = as.numeric(as.character(base::cut(decimalLongitude, breaks=lon_brks, labels=lon_grid))), 
      lat_grid = as.numeric(as.character(base::cut(decimalLatitude, breaks=lat_brks, labels=lat_grid))),
      depth_grid = as.numeric(as.character(base::cut(depth, breaks=depth_brks, labels=depth_grid)))
    )
  
  # group by cell to calculate the number of occurrences occurring on that cell
  
  grouped_occ <- SP %>% 
    group_by(lon_grid, lat_grid, depth_grid, F_AREA) %>%
    dplyr::summarise(n = n()) %>% 
    mutate(status=1) %>% 
    mutate (SP_id = species$SP[i]) %>% 
    mutate (SN = species$SN[i])
  
  # complete the information in the species object
  
  species$n_grouped[i] <- nrow(grouped_occ) # number of occurrences after removing outliers
  
  # save files after removing outliers
  
  write.csv(species, file=here::here ("data","derived_data", "02_species_occurrences", "species_FINAL_list_grouped.csv"))
  
  # give again the proper name
  
  assign(species$SP[i], grouped_occ)
  
  # save files after removing outliers
  
  save (list=species$SP[i], file = here::here ("data","derived_data", "02_species_occurrences", "03_species_occurrences_aggregated", paste0(species$SP[i], "new.RData")))  
  
  print(species$SP[i])
  
  
  # we save a plot for each species
  pdf(here::here ("data","derived_data", "02_species_occurrences", "03_species_occurrences_aggregated","plots_grouped", paste0("grouped_",species$SP[i],".pdf")))
  
  # map of occurrences
  
  p1 <- p0 + 
    geom_point(data=SP, aes(x=decimalLongitude, y=decimalLatitude), col="red", cex=0.3)+
    ggtitle(paste0("Raw data ", " n=",nrow(SP)))
  
  p2 <- p0 + 
    geom_point(data=grouped_occ, aes(x=lon_grid, y=lat_grid), col="red", cex=0.3)+
    ggtitle(paste0("Gridded data ", " n=",nrow(grouped_occ)))
  
  grid.arrange(p1, p2, nrow=1,
               top = textGrob(paste0(species$SP[i],": ", species$SN[i]),gp=gpar(fontsize=20,font=3)))
  
  # Plotting species occurrence points along the depth gradient
  
  p1 <- ggplot(SP, aes(x = decimalLatitude, y = negative_depth)) + 
    geom_pointdensity(alpha=0.3) + 
    scale_color_viridis_c()+
    ggtitle("Raw data")
  p2 <- ggplot(grouped_occ, aes(x = lat_grid, y = -depth_grid)) + 
    geom_pointdensity(alpha=0.3) + 
    scale_color_viridis_c()+
    ggtitle("Gridded data")
  grid.arrange(p1, p2, nrow=1)
  
  dev.off()
  
}
```
