==> scripts/02_species_occurrences/06_selecting_species_for_modelling.Rmd <==
---
title: "Selecting species for modelling"
author: "Mireia Valle"
date: "27/7/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warming = FALSE)
```

### Loading required libraries

```{r}
library (here) #enable easy file referencing
library(tidyverse)
```

### Load species list

```{r}
species <- read.csv (here::here ("data","derived_data", "02_species_occurrences", "species_FINAL_list_grouped.csv"),  sep = ",")

species <- species %>% 
  dplyr::select (-X.1, -X)
```

### Filtering the list

#### subseting
```{r}
tunas <- species %>% 
  filter (taxon_key %in% c("600147", 
                        "600146",
                        "600143",
                        "600142",
                        "600107"))

anchovy <- species %>% 
  filter (taxon_key == "600066")

fish <-  species %>% 
  filter(!(taxon_key %in% c("600147", 
                        "600146",
                        "600143",
                        "600142",
                        "600107", 
                        "600066")))
```

#### selecting those fish species with high number of grouped occurrences (> 1500) and top captured
```{r}
high_occ_top_captured_fish<- fish %>% 
  #filtering species by number of grouped occurrences
  dplyr::filter (n_grouped > 1500) %>% 
  top_n(24, mean_no_bs)
```

#### join all species in one list
```{r}
species_for_modelling <- rbind(tunas, anchovy, high_occ_top_captured_fish)

species_for_modelling <- species_for_modelling %>% 
  dplyr::select(-c(numf, num_abs_f, n_occ))
```

#### saving the list
```{r}
write.csv(species_for_modelling,file=here::here ("data","derived_data","outputs_for_modelling", "species_for_modelling_new.csv"))
```
